OpenEMIS is NOT open source
The source code of OpenEMIS is not in the Google code repository, nor is it anywhere else online.  OpenEMIS core is not available under any license recognized by the open source foundation.  

See their FAQ:

Where can I download the software?
The software is currently available via a shared Demo site. In terms of implementation we offer a range of services to host the software or install and configure the software on a local server for approved countries. For more information please see the OpenEMIS License Agreements.  Strangely their agreements have both an open source license there and  a confidentiality / non disclosure agreement.
There are no files on the Google code site SVN which was not updated in the last years.  No result to searching for their source code. They seem to want people to sign an expression of interest first: https://www.openemis.org/files/resources/OpenEMIS_EOI_en.pdf .

I spoke with someone from their team at a conference: They spoke about maintaining the quality of the code; as if making it under a GPL  license / putting the source code online required them to accept every contribution ever made.